Stop copying systemd libraries from the host.
This hack doesn't work on e.g. CentOS 7 where systemd links to versions
of libgcrypt etc. that are not available in the container.

This hack was used to ensure the systemd libraries in the container
could access the journald files from the host, which could use either XZ
or LZ4 compression, and the fluentd-gcp-image only had support for XZ.

This is no longer the case, since the base image debian-base:0.3 is now
based on Debian Stretch (and not Debian Jessie) which supports both XZ
and LZ4, so this hack is no longer required.

Also: Properly quote "$@" in fluentd command-line. That is the correct
usage of $@.  And exec fluentd binary, so the shell is no longer in the
way. There's really no reason to keep a shell running inside the
container...

Bump version of fluentd-gcp container to 2.0.16 in order to release
these changes as a new version.