Fix memory leak when collecting prometheus metrics.
For each prometheus metrics request, a separate http.Client object is created. After the request is completed, the connections created during its operation remain for the possibility of their reuse. Which leads to the accumulation of goroutine and memory leak.
Instead of creating http.Client for each request, we can reuse once created object, similar to how it was done in net/http package.