Missing Scopes in call to DefaultTokenSource()
One invocation of google.DefaultTokenSource() is missing the scope
parameter. This results in that particular authentication flow failing
to fetch a token due to an empty set of scopes in the request. This PR
sets a wide scope matching another invocation of a similar method. IAM
should restrict perms further.

fixes: #401